agent-gt8004/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
google-generativeai>=0.8.0
httpx>=0.27.0
fastmcp>=2.0
gt8004-sdk[mcp] @ git+https://github.com/vataops/gt8004-sdk.git
